Create MXN_spirat_mat.cpp
Problem Statement :
Write a program that reads an MxN matrix A and prints its elements in spiral order. You should start from the element in the 0th row and 0th column in the matrix and proceed in a spiral order.

Input Format :
First line contains two integers M and N separated by whitespace. The next M lines contain elements of matrix A, starting with the topmost row. Within each row, the elements are given from left to right. (1<=(M,N)<=1000)

Output Format :
A single line consisting of the elements of the matrix printed in a spiral order. All the elements should be separated by whitespace.

Sample Input :
4 4
1 2 3 4
5 6 7 8
9 10 11 12
13 14 15 16

Sample Output :
1 2 3 4 8 12 16 15 14 13 9 5 6 7 11 10